requirements.txt
Flask==2.3.3
flask-sock==0.6.0
requests==2.31.0
websockets==11.0.3
numpy==1.24.3
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
protobuf==4.24.4
grpcio==1.57.0
grpcio-status==1.57.0
Werkzeug==2.3.7
gunicorn==21.2.0
gevent==22.10.1
eventlet==0.33.3